grammar: Hesitate to test s/fdef grammar
It would be nice if grammars threw a helpful error message when given rules with
errors. I was going to charge in with some `fdef` specs for the `grammar`
function but, what's the plan? Instrument rules before generating a grammar?
Write a generator for rules? It's significantly deeper into spec. So, I will
switch gears and do specific grammars and/or corpora next.