magic, rules: Add to grammar, and s/explain it in test
My grammar constructor doesn't conform its data yet. Since it's easy to make
mistakes when adding new terms to a grammar, I put in a line to explain it, so
that when my test fails I can look at the output to figure out what token I've
forgotten to surround in hashes.

While doing that, I also realized there's a subtler spec to write: it ought to be
the case that any expander appears as a key in the grammar. Or, at least there
should be a warning, so that the user can assert, yeah, that expander gets
defined in a to-be-merged rule.